LLMOps/Intro2GenAI/Proj1-Local_LLM_Host/requirements.txt
openai
python-dotenv
fastapi
uvicorn[standard]
tiktoken
qdrant-client
sentence-transformers
pandas
pydantic



